We've managed to get the modal to work
In order for the modal to work, we had to use FlowbiteInstances.getInstance('Modal', 'element')

The reason being, if we use new Modal with import modal, flowbite just creates a new modal without checking if there's one already for that element,
and that would cause troubles and bugs that we can't fix from our js.

A second note, we can't use modalEl as a second urgument for flowbiteinstances, it has to be the id itself or an element with getElementById but since we are using query selector, that made it unusable.

WIP with getting the data and making use of it